--- tests/level3_test.go
@@ -1,1 +1,0 @@
-package caf_tests
